requirements.txt
streamlit~=1.32.2
streamlit-elements~=0.1.0
pandas~=2.2.1
numpy~=1.26.4
pyyaml~=6.0.1
pillow~=10.2.0
scipy~=1.12.0
matplotlib~=3.8.3
opencv-python~=4.9.0.80